requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.97
botocore==1.34.97
Brotli==1.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.2
cssselect2==0.7.0
dj-database-url==2.1.0
Django==4.2.10
django-cors-headers==4.3.1
django-crontab==0.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.48.1
gunicorn==21.2.0
html5lib==1.1
idna==3.6
jmespath==1.0.1
kombu==5.3.7
lxml==5.1.0
num2words==0.5.13
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
pdfkit==1.0.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.0.1
PyPDF2==3.0.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.4
reportlab==4.0.9
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.0
webencodings==0.5.1
xhtml2pdf==0.2.15
zopfli==0.2.3
